pyproject.toml
[project]
name = "bank-analyzer"
version = "0.1.0"
description = ""
authors = [
    {name = "Ian Kiyoshi Kobayashi",email = "contato.iankyoo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.45.0,<0.46.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
]

[tool.poetry]
packages = [{include = "bank_analyzer", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.15.11,<0.16.0)",
    "mypy (>=1.20.2,<2.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true

[tool.taskipy.tasks]
lint = "ruff check ."
format = "ruff format ."
typecheck = "mypy src"
test = "pytest tests/"